Fix app-drawer-layout resetLayout logic
When using the app-drawer-layout on my screen the drawer is opened by default.
This appeared to be in conflict the expected default for `opened`.
After digging around I found what appeared to be a logic error.

Drawer is initially closed when the page is loaded

Drawer is initially open and does not have a transparent backdrop