fix(template): prevent drawer from being swept open on larger screens
Fixes #434

This fixes #434 by binding the `swipe-open` attribute of `app-drawer` to the `smallScreen` media query. As a result, `swipe-open` is only enabled when the viewport is <= 400px wide.

I see that the [contributing guidelines](https://github.com/PolymerElements/app-layout/blob/master/CONTRIBUTING.md) state that tests must be added for bugfixes, but it seems that these templates do not have tests.